header and top navigation
styling for all screen size devices is finished. I changed the size of the font in the dropdown menu. I also changed it's margins and paddings to make the dropdown menu to fit inside the medium sized screen size (as well as the maximum sized one). I tested if everything is displayed properly.