div.content and dropdown menu issue
I discovered that the position: absolute; of my div.content was affecting the dropdown menu (above) and that dropdown menu was not working. Also, when I opened my website in Edge, it was the same issue but also the div.content itself was displayed too high and partly covered the top menu. I tried to find out what might be the issue and when I disabled the position: absolute; from that div, my dropdown menu worked properly but my hero text was displayed below the image. To fix this issue, I decided to give the position: absolute; property to the child div of div.content and add the negative margin top to it. That fixed all issues.