footer
I placed my footer elements inside a container div. I changed all margins and padding to make the whole content that is inside it be displayed in a row on big screen devices but still in columns on the small ones. I also had to make some minor styling changes to this section and removed unused and repetitive styling.