side menu styling
I added more styling to my side menu. I deleted and changed divs classes for my side menu. I used a wrapper to wrap the entire body and added a style to it. I created a class for my side menu that will be added when the menu is active.