Side menu
I made some changes to my index.html. I moved the side menu elements to a div above my header and added the onClick() element to it. I wrote a JavaScript function that adds or removes classes from side menu divs and whole body elements and moves the website's content to the left when the button is clicked and the menu opens. The only issue there is that one class that I want to add to the body element (that moves the content) is undefined so it needs to be checked and resolved.