pop-up cookies policy
In HTML, I added a second paragraph to my modal div, I added extra margin to the first paragraph. I added a general style to "row". I wrote a function for my modal in JavaScript and modified the function for my side menu.  I also wrote a function that stick my header to the top.