About section and footer
I made some changes in the middle section of the website and in the footer. I changed margins around the "Netmatters" section of the website, the one that says about Netmatters. I also added extra margins to the "News" elements. In the footer, I centred all elements and removed the styling that should be applied only on bigger screen devices.